requirements.txt
#use python version 3.10.4
#conda install -c conda-forge jsonnet
aiohttp==3.8.1
aiosignal==1.2.0
allennlp==2.9.3
allennlp-models==2.9.3
appnope==0.1.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
blis==0.7.7
boto3==1.24.2
botocore==1.27.2
cached-path==1.1.2
cachetools==5.2.0
catalogue==2.0.7
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.4
conllu==4.4.1
cymem==2.0.6
datasets==2.2.2
decorator==5.1.1
dill==0.3.4
distlib==0.3.4
docker-pycreds==0.4.0
executing==0.8.3
fairscale==0.4.6
filelock==3.6.0
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
gitdb==4.0.9
gitpython==3.1.27
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
h5py==3.7.0
huggingface-hub==0.5.1
identify==2.5.1
idna==3.3
iniconfig==1.1.1
ipython==8.4.0
jedi==0.18.1
jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
jsondiff==1.3.0
langcodes==3.3.0
lmdb==1.3.0
markupsafe==2.1.1
matplotlib-inline==0.1.3
mead-baseline==2.4.2
mead-layers==2.4.2
more-itertools==8.13.0
multidict==6.0.2
multiprocess==0.70.12.2
murmurhash==1.0.7
mypy-extensions==0.4.3
nltk==3.7
nodeenv==1.6.0
numexpr==2.8.1
numpy==1.22.4
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathtools==0.1.2
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.17.0
preshed==3.0.6
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-rouge==1.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pygments==2.12.0
pyhocon==0.3.58
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-rapidjson==1.6
pytz==2022.1
pyyaml==6.0
regex==2022.6.2
requests==2.27.1
responses==0.18.0
rsa==4.8
s3transfer==0.6.0
sacremoses==0.0.45
scikit-learn==1.1.1
scipy==1.8.1
sentencepiece==0.1.96
sentry-sdk==1.5.12
seqeval==1.2.2
serpyco==1.3.7
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.2.0
tensorboardx==2.5
termcolor==1.1.0
thinc==8.0.17
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
traitlets==5.2.2.post1
transformers==4.18.0
typer==0.4.1
typing-extensions==4.2.0
typing-inspect==0.7.1
urllib3==1.26.9
virtualenv==20.14.1
wandb==0.12.17
wasabi==0.9.1
wcwidth==0.2.5
word2number==1.1
xxhash==3.0.0
yarl==1.7.2
